Consolidate and optimize test suite, reducing redundant files and improving organization
